Add files via upload
v0.m is the first of many v.m files we will create. New ones should be committed after it is established that they work. So far I tested this v0 just in octave, but we should test it in matlab too and make sure there is cross compatibility. variableDefinition.m is a script that holds all the variables names, etc. I've coded these files the way I have to attempt to make it easy to add new entities/ODEs into the model.